backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "uvicorn>=0.31.0",
]

[tool.uv]
dev-dependencies = [
    "black>=24.8.0",
    "isort>=5.13.2",
]
